fix to combine fetch and helper function to one file
