+++ src/applications/vaos/appointment-list/pages/RequestedAppointmentDetailsPage/RequestedAppointmentDetailsPage.unit.spec.jsx
@@ -3,18 +3,18 @@ import { fireEvent, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import { expect } from 'chai';
+import { addDays, subDays } from 'date-fns';
 import MockDate from 'mockdate';
-import moment from 'moment';
 import React from 'react';
 import { AppointmentList } from '../..';
 import MockAppointmentResponse from '../../../tests/e2e/fixtures/MockAppointmentResponse';
 import MockFacilityResponse from '../../../tests/e2e/fixtures/MockFacilityResponse';
-import { mockFacilityFetch } from '../../../tests/mocks/fetch';
 import {
   mockAppointmentApi,
-  mockAppointmentUpdateApi,
   mockAppointmentsApi,
-} from '../../../tests/mocks/helpers';
+  mockAppointmentUpdateApi,
+  mockFacilityApi,
+} from '../../../tests/mocks/mockApis';
 import {
   createTestStore,
   getTestDate,
@@ -49,12 +49,8 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
       status: APPOINTMENT_STATUS.proposed,
     });
     mockAppointmentsApi({
-      end: moment()
-        .add(2, 'day')
-        .format('YYYY-MM-DD'),
-      start: moment()
-        .subtract(120, 'days')
-        .format('YYYY-MM-DD'),
+      end: addDays(new Date(), 2),
+      start: subDays(new Date(), 120),
       statuses: ['proposed', 'cancelled'],
       response: [response],
     });
@@ -220,8 +216,8 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
 
     mockAppointmentApi({ response });
     mockAppointmentUpdateApi({ response: canceledResponse });
-    mockFacilityFetch({
-      facility: new MockFacilityResponse({ id: '983' }),
+    mockFacilityApi({
+      response: new MockFacilityResponse({ id: '983' }),
     });
 
     // Act
@@ -262,8 +258,8 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
 
     mockAppointmentApi({ response });
     mockAppointmentUpdateApi({ response: canceledResponse });
-    mockFacilityFetch({
-      facility: new MockFacilityResponse({ id: '983' }),
+    mockFacilityApi({
+      response: new MockFacilityResponse({ id: '983' }),
     });
 
     // Act
@@ -311,12 +307,8 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
         pending: true,
       });
       mockAppointmentsApi({
-        end: moment()
-          .add(2, 'day')
-          .format('YYYY-MM-DD'),
-        start: moment()
-          .subtract(120, 'days')
-          .format('YYYY-MM-DD'),
+        end: addDays(new Date(), 2),
+        start: subDays(new Date(), 120),
         statuses: ['proposed', 'cancelled'],
         response: [response],
       });
@@ -358,18 +350,14 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
 
       mockAppointmentApi({ response });
       mockAppointmentsApi({
-        end: moment()
-          .add(1, 'day')
-          .format('YYYY-MM-DD'),
-        start: moment()
-          .subtract(120, 'days')
-          .format('YYYY-MM-DD'),
+        end: addDays(new Date(), 1),
+        start: subDays(new Date(), 120),
         statuses: ['proposed', 'cancelled'],
         response: [response],
       });
       mockAppointmentUpdateApi({ response: canceledResponse });
-      mockFacilityFetch({
-        facility: new MockFacilityResponse({ id: '983' }),
+      mockFacilityApi({
+        response: new MockFacilityResponse({ id: '983' }),
       });
 
       // Act
@@ -432,18 +420,14 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
       mockAppointmentApi({ response });
 
       mockAppointmentsApi({
-        start: moment()
-          .subtract(120, 'days')
-          .format('YYYY-MM-DD'),
-        end: moment()
-          .add(1, 'day')
-          .format('YYYY-MM-DD'),
+        start: subDays(new Date(), 120),
+        end: addDays(new Date(), 1),
         statuses: ['proposed', 'cancelled'],
         response: [],
       });
       mockAppointmentUpdateApi({ response: canceledResponse });
-      mockFacilityFetch({
-        facility: new MockFacilityResponse({ id: '983' }),
+      mockFacilityApi({
+        response: new MockFacilityResponse({ id: '983' }),
       });
 
       // Act
@@ -509,12 +493,8 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
       mockAppointmentApi({ response });
 
       mockAppointmentsApi({
-        start: moment()
-          .subtract(120, 'days')
-          .format('YYYY-MM-DD'),
-        end: moment()
-          .add(1, 'day')
-          .format('YYYY-MM-DD'),
+        start: subDays(new Date(), 120),
+        end: addDays(new Date(), 1),
         statuses: ['proposed', 'cancelled'],
         response: [],
       });
@@ -522,8 +502,8 @@ describe('VAOS Page: RequestedAppointmentDetailsPage', () => {
         response: canceledResponse,
         responseCode: 500,
       });
-      mockFacilityFetch({
-        facility: new MockFacilityResponse({ id: '983' }),
+      mockFacilityApi({
+        response: new MockFacilityResponse({ id: '983' }),
       });
 
       // Act
