+++ src/applications/vaos/appointment-list/pages/RequestedAppointmentsPage/RequestedAppointmentsPage.unit.spec.jsx
@@ -5,13 +5,13 @@ import { within } from '@testing-library/dom';
 import { format, addDays, subDays } from 'date-fns';
 import React from 'react';
 import reducers from '../../../redux/reducer';
-import { mockVAOSAppointmentsFetch } from '../../../tests/mocks/helpers';
 import { getVAOSRequestMock } from '../../../tests/mocks/mock';
 import {
   getTestDate,
   renderWithStoreAndRouter,
 } from '../../../tests/mocks/setup';
 import RequestedAppointmentsPage from './RequestedAppointmentsPage';
+import { mockVAOSAppointmentsFetch } from '../../../tests/mocks/mockApis';
 
 const initialStateVAOSService = {
   featureToggles: {
