+++ src/applications/vaos/covid-19-vaccine/components/ClinicChoicePage.unit.spec.jsx
@@ -1,19 +1,19 @@
-import React from 'react';
-import { expect } from 'chai';
+import { mockFetch } from '@department-of-veterans-affairs/platform-testing/helpers';
 import { waitFor } from '@testing-library/dom';
 import { cleanup } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { mockFetch } from '@department-of-veterans-affairs/platform-testing/helpers';
+import { expect } from 'chai';
+import React from 'react';
 import {
   createTestStore,
   renderWithStoreAndRouter,
   setVaccineFacility,
 } from '../../tests/mocks/setup';
 
-import ClinicChoicePage from './ClinicChoicePage';
-import { TYPE_OF_CARE_ID } from '../utils';
-import { mockEligibilityFetches } from '../../tests/mocks/fetch';
 import { createMockClinic } from '../../tests/mocks/data';
+import { mockEligibilityFetches } from '../../tests/mocks/mockApis';
+import { TYPE_OF_CARE_ID } from '../utils';
+import ClinicChoicePage from './ClinicChoicePage';
 
 const initialState = {
   featureToggles: {
