+++ src/applications/vaos/new-appointment/components/ClinicChoicePage/index.unit.spec.jsx
@@ -1,9 +1,9 @@
-import React from 'react';
-import { expect } from 'chai';
+import { mockFetch } from '@department-of-veterans-affairs/platform-testing/helpers';
 import { waitFor } from '@testing-library/dom';
 import { cleanup } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { mockFetch } from '@department-of-veterans-affairs/platform-testing/helpers';
+import { expect } from 'chai';
+import React from 'react';
 import {
   createTestStore,
   renderWithStoreAndRouter,
@@ -12,9 +12,9 @@ import {
 } from '../../../tests/mocks/setup';
 
 import ClinicChoicePage from '.';
-import { getV2ClinicMock } from '../../../tests/mocks/mock';
 import { createMockCheyenneFacility } from '../../../tests/mocks/data';
-import { mockEligibilityFetches } from '../../../tests/mocks/fetch';
+import { getV2ClinicMock } from '../../../tests/mocks/mock';
+import { mockEligibilityFetches } from '../../../tests/mocks/mockApis';
 
 const initialState = {
   featureToggles: {
