+++ src/applications/vaos/new-appointment/components/ReviewPage/index.va-request.unit.spec.jsx
@@ -1,26 +1,28 @@
 /* eslint-disable camelcase */
-import React from 'react';
 import { expect } from 'chai';
+import React from 'react';
 
-import userEvent from '@testing-library/user-event';
 import { waitFor, within } from '@testing-library/dom';
+import userEvent from '@testing-library/user-event';
 
 import {
-  setFetchJSONFailure,
   mockFetch,
+  setFetchJSONFailure,
 } from '@department-of-veterans-affairs/platform-testing/helpers';
 import environment from '@department-of-veterans-affairs/platform-utilities/environment';
 
-import { FACILITY_TYPES } from '../../../utils/constants';
 import {
   createTestStore,
   renderWithStoreAndRouter,
 } from '../../../tests/mocks/setup';
+import { FACILITY_TYPES } from '../../../utils/constants';
 
 import ReviewPage from '.';
-import { mockAppointmentSubmit } from '../../../tests/mocks/helpers';
 import { createMockCheyenneFacility } from '../../../tests/mocks/data';
-import { mockFacilityFetch } from '../../../tests/mocks/fetch';
+import {
+  mockAppointmentSubmit,
+  mockFacilityApi,
+} from '../../../tests/mocks/mockApis';
 
 const initialState = {
   featureToggles: {
@@ -216,8 +218,8 @@ describe('VAOS Page: ReviewPage VA request with VAOS service', () => {
 
   it('should show error message on failure', async () => {
     store = createTestStore(defaultState);
-    mockFacilityFetch({
-      facility: createMockCheyenneFacility({}),
+    mockFacilityApi({
+      response: createMockCheyenneFacility({}),
     });
     setFetchJSONFailure(
       global.fetch.withArgs(`${environment.API_URL}/vaos/v2/appointments`),
